requirements.txt
pyspark==3.5.3
pymongo==4.6.3
cassandra-driver==3.29.2
redis==5.0.8
psycopg2-binary==2.9.9
clickhouse-connect==0.6.7
neo4j==5.19.0
PyYAML==6.0.1
